feat: implemented legacy TradeConfirm for swapi
