+++ src/components/MultiHopTrade/components/TradeInput/components/SwapperIcon/0x-icon.png
