+++ src/components/MultiHopTrade/components/TradeInput/components/SwapperIcon/1inch-icon.png
