+++ src/components/MultiHopTrade/components/TradeInput/components/SwapperIcon/cow-icon.png
