+++ src/components/MultiHopTrade/components/TradeInput/components/SwapperIcon/lifi-icon.png
