+++ src/components/MultiHopTrade/components/TradeInput/components/SwapperIcon/osmosis-icon.png
