+++ src/components/MultiHopTrade/components/TradeInput/components/SwapperIcon/thorchain-icon.png
